// state/Header/header.action-types.ts
export enum ActionTypes {
    HEADER_UPLOAD_REQUEST = 'HEADER_UPLOAD_REQUEST',
    HEADER_UPLOAD_SUCCESS = 'HEADER_UPLOAD_SUCCESS',
    HEADER_UPLOAD_FAIL = 'HEADER_UPLOAD_FAIL',

    HEADER_FETCH_REQUEST = 'HEADER_FETCH_REQUEST',
    HEADER_FETCH_SUCCESS = 'HEADER_FETCH_SUCCESS',
    HEADER_FETCH_FAIL = 'HEADER_FETCH_FAIL',

    HEADER_DELETE_REQUEST = 'HEADER_DELETE_REQUEST',
    HEADER_DELETE_SUCCESS = 'HEADER_DELETE_SUCCESS',
    HEADER_DELETE_FAIL = 'HEADER_DELETE_FAIL',
    
    HEADER_RESET = 'HEADER_RESET',
  }